+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - print jobs can specify a media-source (paper tray) via the `media-col` attribute
   ([issue #21](https://github.com/harwey/cups4j/issues/21))
+- `mediaSource(..)` added to `PrintJob.Builder` as suggested in [issue #21](https://github.com/harwey/cups4j/issues/21)
 
 
 ## [0.8.2] (02-Jun-2026)